Color-code the windows in status bar
Show legacy admin windows with a green background, API windows with a
cyan background. This is to better indicate that you need to use a
special key binding to jump to some.

tmate does not support the `==:`, and especially not the `>:`,
comparison operators, but it does support external scripts. Add a quick
`tmux-colors` script to choose between the two.